# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NurturetheSpectrum.com API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Listing {
  id: number;
  name: string;
  slug: string;
  type: string;
  city: string;
  county: string;
  specializations: string[];
  certifications: string[];
  /** @nullable */
  website?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  hourlyRate?: string | null;
  /** @nullable */
  yearsExperience?: number | null;
  description: string;
  /** @nullable */
  logoUrl?: string | null;
  tier: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface ListingsResponse {
  listings: Listing[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CountStat {
  label: string;
  count: number;
}

export interface ListingStats {
  byCounty: CountStat[];
  byType: CountStat[];
  byTier: CountStat[];
  total: number;
}

export interface SubmitListingBody {
  name: string;
  type: string;
  city: string;
  county: string;
  specializations?: string[];
  certifications?: string[];
  website?: string;
  email: string;
  phone?: string;
  hourlyRate?: string;
  /** @nullable */
  yearsExperience?: number | null;
  description: string;
  logoUrl?: string;
}

export interface SubmitListingResponse {
  success: boolean;
  message: string;
}

export interface AdminListingBody {
  name?: string;
  type?: string;
  city?: string;
  county?: string;
  specializations?: string[];
  certifications?: string[];
  website?: string;
  email?: string;
  phone?: string;
  hourlyRate?: string;
  /** @nullable */
  yearsExperience?: number | null;
  description?: string;
  logoUrl?: string;
  tier?: string;
  status?: string;
}

export interface ContactBody {
  name: string;
  email: string;
  message: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message: string;
}

export interface AdminMeResponse {
  authenticated: boolean;
}

export interface DashboardStats {
  totalListings: number;
  pendingCount: number;
  publishedCount: number;
  inactiveCount: number;
  freeCount: number;
  featuredCount: number;
  verifiedCount: number;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetListingsParams = {
  county?: string;
  type?: string;
  specializations?: string;
  certifications?: string;
  page?: number;
  limit?: number;
};

export type AdminGetListingsParams = {
  status?: string;
  page?: number;
  limit?: number;
};