# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NurturetheSpectrum.com API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: adminGetMe
      tags: [admin]
      summary: Check admin session
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/listings:
    get:
      operationId: adminGetListings
      tags: [admin]
      summary: Get all listings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All listings for admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsResponse"
    post:
      operationId: adminCreateListing
      tags: [admin]
      summary: Create a new listing (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminListingBody"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/listings/{id}:
    patch:
      operationId: adminUpdateListing
      tags: [admin]
      summary: Update a listing (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminListingBody"
      responses:
        "200":
          description: Listing updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteListing
      tags: [admin]
      summary: Delete a listing (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/listings/{id}/approve:
    post:
      operationId: adminApproveListing
      tags: [admin]
      summary: Approve a pending listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        type:
          type: string
        city:
          type: string
        county:
          type: string
        specializations:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hourlyRate:
          type: ["string", "null"]
        yearsExperience:
          type: ["integer", "null"]
        description:
          type: string
        logoUrl:
          type: ["string", "null"]
        tier:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - type
        - city
        - county
        - specializations
        - certifications
        - description
        - tier
        - status
        - createdAt
        - updatedAt
    ListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - listings
        - total
        - page
        - totalPages
...
    AdminListingBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        city:
          type: string
        county:
          type: string
        specializations:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        website:
          type: string
        email:
          type: string
        phone:
          type: string
        hourlyRate:
          type: string
        yearsExperience:
          type: ["integer", "null"]
        description:
          type: string
        logoUrl:
          type: string
        tier:
          type: string
        status:
          type: string
...
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    AdminMeResponse:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error