# lib/api-zod/src/generated/types/adminListingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NurturetheSpectrum.com API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminListingBody {
  name?: string;
  type?: string;
  city?: string;
  county?: string;
  specializations?: string[];
  certifications?: string[];
  website?: string;
  email?: string;
  phone?: string;
  hourlyRate?: string;
  /** @nullable */
  yearsExperience?: number | null;
  description?: string;
  logoUrl?: string;
  tier?: string;
  status?: string;
}