# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NurturetheSpectrum.com API
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  id: number;
  name: string;
  slug: string;
  type: string;
  city: string;
  county: string;
  specializations: string[];
  certifications: string[];
  /** @nullable */
  website?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  hourlyRate?: string | null;
  /** @nullable */
  yearsExperience?: number | null;
  description: string;
  /** @nullable */
  logoUrl?: string | null;
  tier: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}