# lib/api-zod/src/generated/types/listingsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NurturetheSpectrum.com API
 * OpenAPI spec version: 0.1.0
 */
import type { Listing } from "./listing";

export interface ListingsResponse {
  listings: Listing[];
  total: number;
  page: number;
  totalPages: number;
}